% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bristle_sample}
\alias{bristle_sample}
\title{Sample Bristle Raw Data}
\format{
A data frame 109 rows and 3 columns:
\describe{
  \item{genus}{char}
  \item{species}{char}
  \item{abundance}{numeric}
}
}
\source{
\url{http://bristlehealth.com/}
}
\usage{
data(bristle_sample)
}
\description{
A typical XLSX file from a Bristle user
}
\keyword{datasets}
