% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristler-package.R
\docType{package}
\name{bristler-package}
\alias{bristler}
\alias{bristler-package}
\title{bristler: Analyze Oral Microbiome Data from Bristle Health}
\description{
Read Bristle-formated oral microbiome data Generate plots and do other analysis.
}
\keyword{internal}
