% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristle_read.R
\name{clip_bristle_table}
\alias{clip_bristle_table}
\title{Copy Clipboard into Tidy Dataframe in Canonical Form}
\usage{
clip_bristle_table()
}
\value{
tidy data frame of Bristle data in canonical form

dataframe
}
\description{
Copy Clipboard into Tidy Dataframe in Canonical Form
}
