% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristle_read.R
\name{combine_bristle_table}
\alias{combine_bristle_table}
\title{Merge Two Bristle Dataframes}
\usage{
combine_bristle_table(table1, table2, label1, label2)
}
\arguments{
\item{table1}{dataframe in Bristle canonical form}

\item{table2}{dataframe in Bristle canonical form}

\item{label1}{label for dataframe}

\item{label2}{label for dataframe}
}
\value{
dataframe
}
\description{
Merge Two Bristle Dataframes
}
