% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristle_phyloseq.R
\name{phyloseqize}
\alias{phyloseqize}
\title{Make a Phyloseq Object from a Bristle Dataframe}
\usage{
phyloseqize(bristle_df, mapfile)
}
\arguments{
\item{bristle_df}{dataframe in canonical Bristle format}

\item{mapfile}{dataframe (see instructions)}
}
\value{
phyloseq object
}
\description{
You'll need to pass both a canonical Bristle dataframe
and a mapfile (see instructions)
}
