% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristle_read.R
\name{plot_bristle_freq}
\alias{plot_bristle_freq}
\title{GGplot Frequency for Sample}
\usage{
plot_bristle_freq(r_table)
}
\arguments{
\item{r_table}{Bristle canonical raw dataframe}
}
\value{
ggplot object
}
\description{
Generate a ggplot of an ordered frequency of the genus in the sample
}
