% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristle_read.R
\name{read_bristle_table}
\alias{read_bristle_table}
\title{Read File Into Tidy Dataframe in Canonical Form}
\usage{
read_bristle_table(
  filepath = system.file("extdata", package = "bristler", "BristleHealthRaw.xlsx")
)
}
\arguments{
\item{filepath}{path to XLSX file}
}
\value{
tidy data frame of Bristle data in canonical form

dataframe
}
\description{
Read File Into Tidy Dataframe in Canonical Form
}
