% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristle_read.R
\name{species_from_species}
\alias{species_from_species}
\title{Species name from species}
\usage{
species_from_species(species_name)
}
\arguments{
\item{species_name}{char Species name}
}
\value{
char species name (as a string)
}
\description{
Return the species of a species. Assumes it's the second word of the full species name.
}
