% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristle_read.R
\name{treemap_of_sample}
\alias{treemap_of_sample}
\title{Dataframe Version of Treemap}
\usage{
treemap_of_sample(sample)
}
\arguments{
\item{sample}{dataframe in Bristle Canonical Form}
}
\value{
dataframe
}
\description{
return a a dataframe version of this treemap
}
