% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bristle_read.R
\name{read_bristle_table}
\alias{read_bristle_table}
\title{Tidy Dataframe in Canonical Form}
\usage{
read_bristle_table(filepath = file.path("data", "BristleHealthRaw.xlsx"))
}
\arguments{
\item{filepath}{path to XLSX file}
}
\value{
tidy data frame of Bristle data in canonical form

dataframe
}
\description{
Tidy Dataframe in Canonical Form
}
